requirements.txt
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
idna==2.9
kiwisolver==1.1.0
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
praw==6.5.1
prawcore==1.0.1
pyparsing==2.4.6
python-dateutil==2.8.1
requests==2.23.0
six==1.14.0
update-checker==0.16
urllib3==1.25.8
websocket-client==0.57.0
